Added Settingsfile & Minor Code cleanup
A settings.txt file is created and filled with
the hard-coded values if none is found.
If one is found its content will overwrite the
hard-coded values.

Checks if the internal version string matches the settings.txt
file before loading the values.
The settings file ignores lines not matching the regex: ".*=.*"